[Cluster] Rename and merge tests around fragmented messages.
